Code review, refactor and styling
Introduces a dev-dependency: phpstan/phpstan. It is a tool for static
analysis of code so that the project code could be maintained without
using deprecated behaviour.

Next on the list of changes is code update. In this commit this is
partial work done, so expect more to come and commits will be squashed.
Some changes introduce comments that mark some functionality that can be
removed but need a discussion on what to do. Other changes introduce
basic way to deal with not returning a 'Response' object in code which
should work (needs testing, but this comment will change in further
commits).

Most of the changes are cosmetic: adding a return statement as the
inherited methods are declared with a return type, documeneted empty
methods that also should return not void, but are empty.