More code clean up for Admin Controllers
Added comments for methods, that should be reworked in order to properly
return their declared object, and parameters, that could probably be
removed safely.

For the rest of methods a simple "return" statements is prepended to
"Illuminate\Foundation\Auth\Access\AuthorizesRequests::authorize" as it
returns the specified object of that method.