Revamp text parsing to content parsing
Created migration, model and factory for Content, which has a
relation with Page and News models as for now. This content for
migration purposes is parsed by Tiptap PHP to JSON. This JSON is
transformed to HTML only on the client. `json_content` may accommodate
not only Tiptap JSON, but custom specifications.